+++ example/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:idle_logout/idle_logout.dart';
 
 import '../../screens/home_screen.dart';
-import '../../screens/other_screen.dart';
+import 'screens/lock_screen.dart';
 
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
@@ -15,23 +15,21 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      navigatorKey: navigatorKey,
-      home: IdleLogout(
-        timeout: const Duration(seconds: 4),
-        isLoggedIn: () => true,
-        isLockedOut: () => false,
-        lockedOutAction: () async {
-          debugPrint('User logged out due to inactivity');
+    return IdleLogout(
+      pauseThreshold: const Duration(seconds: 15),
+      timeout: const Duration(seconds: 10),
+      isLoggedIn: () => true,
+      isLockedOut: () => false,
+      lockedOutAction: () async {
+        debugPrint('User logged out due to inactivity');
 
-          navigatorKey.currentState?.pushReplacement(
-            MaterialPageRoute<void>(
-              builder: (BuildContext context) => const OtherScreen(),
-            ),
-          );
-        },
-        child: const HomeScreen(),
-      ),
+        navigatorKey.currentState?.pushReplacement(
+          MaterialPageRoute<void>(
+            builder: (BuildContext context) => const LockScreen(),
+          ),
+        );
+      },
+      child: MaterialApp(navigatorKey: navigatorKey, home: const HomeScreen()),
     );
   }
 }